Added Project entity, and minor changes.
Added project entity, and management implementations (create/delete
etc). to it.

Furthermore added new interface "Renameable". Via this interface a
ProejctElement can sign, that it can be renamed. In connection to this
interface the name handling on teh DefaultProjectNode is changed.
ProjectElement not has properties, and fires property change events. The
default node takes it's displayname from this property. The existing
project elements are updated to set this property on initialization.